move solana_program::stake_history contents into solana_program::sysvar
